Piece class no longer tracks piece position
This was done for a couple reasons. The first reason is that this was redundant,
as the position was already tracked in the GameEngine's _board property.

Secondly, as I've been planning how to handle checking for checks, I determined
that the best solution will involve creating a "preview" board for what happens
after a move, then checking if any pieces are attacking the king (or whatever piece
can get checked). However, a lot of code relies on a pieces internal position, meaning
I'd have to create brand new pieces for this new board. Alternatively, if the position
is not stored in the piece, this becomes a lot easier.

Finally, the position was only used in a limited number of places, so it was pretty easy
to remove. However, this may be more difficult later, so I wanted to tackle it now.