Refactored move functions to take RectangularBoard as input
The only GameEngine functions used were ones related to the board.
Therefore, it's simpler to remove the middleman and just pass a board
directly.

This also de-couples the move functions from the current board state, which
will make it easier to do things such as checking if a potential move puts
a piece in check.